Refactor StatusTag
refactor: StatusTag is now a component
